refactor: Refactor into Header component
